Refactored code from hero.cs into two new scripts, character.cs and player.cs that utilize inheritance. hero.cs has been deleted as the functionality is now is character.cs and player.cs. The "hero" now contains the player.cs script instead of hero.cs
Currently, the player acts exactly the same as it did before.

Details:
charatcer.cs contains an abstract class with some abstract methods that must be implemented in child classes.
player.cs is a child class of character.cs. In the future when we implement enemies, each enemy will be a
child class of character.cs. This means that enemies will have the same basic implementation as the character.
The difference will be how they get input and how they move. The player will get input from the keyboard while
the enemies will get input through some other method that will determine how to move.